combined forestflourisher and forestflourisherskill; no longer necessary for them to be separate
